requirements.txt
line-bot-sdk>=3.0.0
httpx>=0.27.0
fastapi>=0.115.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
